Organize code into multiple files. Make crop selection more intuitive. Add download buttons. Fix bug with hiding cursor.
