+++ src/dev/colors.ts
@@ -1,0 +1,8 @@
+const COLORS = {
+  bg: '#2f0e6c',
+  shadow: 'rgba(47,14,108,60%)',
+  text: '#f5f7fa',
+  selected: '#8fd5ec',
+  hover: '#dde8f8',
+};
+export default COLORS;
